refactor: PathFavoritesEntity 클래스 위치 path로 분리
<description>
<issue-number>
